Fix checkpoint restoration;
All options for which context switch needs to be taken have to be
blocked by _DFS_traversed indicating they're already taken, /except/ for
the thread we want to hop to next.

So, mark all _DFS_traversed elements except that as true, then restore
the correct contents after we've reached the next context switch point.