Don't allow an operand field to be uninitialized
The arguments to this function should always be an irep with one operand
that's a pointer. If it doesn't, the fall through from the modified if
statement will leave operands[0] uninitialized.

So instead, assert that this is the case. If it isn't then we'll abort,
rather than end up using uninitialized data.