Silence a set of compiler warnings
These are largely related to either not returning a return value along a
particular program path and the potential for uninitialized variables.
For most of the uninitialized variable situations it's due to an
ifelse/ifelse set of statements not being ended with an else.

For those I've bunged assert(false) in to show up the fact we've found
something unimplemented. There's no point trying to continue from such a
situation.