Congeal convert_{add,sub} into same function
They both do almost exactly the same thing, and congealing leads to no
extra regression tests failing. Notably a clause catching cases where
the operands to + typecasts, however this appears to be pointer related,
and thus a target for this refactor.

And anyway it wasn't duplicated for converting -, so it still wouldn't
have been correct.