Make Z3 conversion functions private
We don't expect anything to extend Z3_convt; we also don't want anything
to call conversion methods unexpectedly.

This is in preperation for fixing the error situation that's going on.
Instead of returning an error bool, or returning a "true" or "false" AST
on error, we should be throwing an error exception instead.