Convert create.*type routines to generate exps
This eliminates a sizeable chunk of conditionals that only serve to
propagate an error flag upwards through the call chain.

It also identifies several portions of code that attempt to handle an
error in type conversion with some extra special cases. Ultimately such
cases should be in the original conversion method, not on some outer
portion of code.