Catch conversion exceptions and print error msg
All useful methods in z3_convt are private, and these are the only two
virtual functions that seem to lead to calls to conversion functions, so
I'm happy to say that these are the only two entry points that one can
generate a conv_error exception from.

Ideally in the future we'll get rid of the "ignoring" facility
altogether and instead abort upon error. Removing portions of the
program like that just isn't valid.