Simplify error handling situations in Z3 backend
As alluded to before: we now never return an error code if there's an
error, instead we throw an exception, which is caught in either set_to
or convert_rest (only two entry points to these portions of code).

This leads to a large amount of code reduction (~250 lines) and
simplification, no further regression tests fail.