Eliminate a duplicated is_dynamic_object
Which one of these gets called depends on where in the expr it is, which
is bad because they also have differing behaviours.

The one deleted doesn't extract the object id from the pointer tuple at
all, so it's significantly wrong. The other compares the current object
with a dynamic number, then inverts the output, so it would appear to be
the wrong logic (which isn't the same in the other method).