Eliminate code duplication in create_array_type
IE: we can let create_type handle the problem of creating the subtype,
there's no need for us to reimplement it in this method.

Much more irritating is the fact that this change is leading to a large
number of sort mismatch errors, so it would appear that the
implementations of create_type and create_array_type were differing in
how the /actually/ created types. Grrrr.

(On the other hand, this is exactly the kind of error that I was looking
 to flush out)