Fix a type fault in create_array_of
This is supposed to be populating an array of pointers with a sample
(blank/nil/whatever) pointer. However, what it was actually doing was
populating it with the offset. Which is an invalid typing situation.

Fix this by using the full pointer object, and not selecting the index
from it.