Fix some behavior of convert_index
I don't understand what was happening in the previous code; the value
was being selected, and then being updated into... what?

New implementation just performs an index, as one would expect. I need
to confirm however that a situation where one dereferences a pointer
with an index is executed-out by the symex step; otherwise there might
need to be some extra implementation here to handle the case of:

  int bees[3];
  int *face = bees;
  int cheese = face[2];

Or suchlike.