Adjust convert_invalid method.
It's non-obvious what "invalid" means. From what I've worked out, it
seems to mean evaluating whether or not the pointer is a pointer to the
"invalid" object identifier. And the invalid object identifier
represents... well, I don't know. Possibly it's /supposed/ to be the
initial values of all pointers.

In any case; Lucas clearly decided to interpret this in a way that tried
to work out whether or not the pointer was /valid/, ie, pointed at null,
etc. Remove this and make the probably-expected comparison.

This leads to a lot of breakage of regression tests. This is the point
where I give up trying to make pointers work, and just start rewriting
stuff.