Simplify convert_identifier
No need to have this huge if/else case to create various types, use
create_type instead.

The one exception that makes a difference appears to be that unsigned
integers need to be assumed-to-be zero or greater. This is so that, if
integer-represented, that they're forced to be non-negative.

I don't know whether this is the only entry point via which negative
integers can enter what's supposed to be unsigned. But at least for
nondeterminism, which becomes a free symbol, I think this is correct.