On second thoughts, update convert_same_object
So, after thinking about it a little, the same-object irep is used in
the dereference switch statements to see whether or not a pointer points
at a particular thing in "memory".

So, if we had two arrays, and were dereferencing to write to one of them
through a pointer, it'd issue a same-object against the existing pointer
and use that to determine the results.

The point here being that there are circumstances where the offset
_isn't_ something we want to make part of the comparison.