Simplify convert_dynamic_object
We shouldn't need to care about what the int width is of the pointer
tuple; it's an internal object, we get to dictate what it is. So,
operate by assuming it's int_width.

Also, remove a duplication of the "return false if no dyn objects"
optimisation. It seems it's always been there, and I just added it a
while ago.