Strip out huge amounts of pointer typecast goo
Essentially, we're running into the problem repeatedly that there's
currently no decent way of representing what happens when you typecast a
pointer into an integer, or an integer into a pointer, and so forth.
Right now there seem to be a large amount of contradictory things
sprinkled over the Z3 backend that convert pointers back and forth,
with no documentation and as far as I can tell no overall purpose.

So to get around this, now aim to put all typecasts /to/ a pointer into
one method, and all typecasts /from/ a pointer into one method. This
should actually allow us to seperate all pointer logic from other goo.

As for how to actually implement said methods, that's for the future.