Don't over-type pointer types.
So right now (and previously in the Z3 backend) whenever we create a
pointer to an object, it gets it's own specially named type.

A partial reason for this is because on occasion the Z3 backend will
spit out a pointer where the pointer object is in fact the symbol for
the object itself, not an identifier. Which breaks absolutely everything
pointerwise. This is one of the things that's going to get stripped out.

More generally, I don't see a reason for making a pointer tuple's type
be tied to what it points at. The operations on the pointer tuple itself
don't change depending on what it points at, and type safety is the job
of whatever code then converts back from a pointer to the object itself.