Simplify convert_pointer_object.
Simplify is not necessary the right word; "delete" may be.

This function contained ~200 lines to select the pointer object from a
pointer, and insisted on performing different thigns for different ireps
when it could have just made the selection. Uugh.

Even better, no extra regression tests fail because of this, although
many of the pointer ones are already going to be failing.