Start implementing ptr distinction functionality
The actual operation of this should be documented elsewhere; however the
idea is to allow dynamic (in SMT) lookup of where a pointers bounds are
in some abstract address space. This array records the mapping between
obj IDs and said address space range. The elements are (start, end) as
integers, indexes are the obj ids as integers.