Implement taking the address of a pointer constant
Currently by generating a dummy symbol for the value and taking the
address of it; it'll be the same object for the same string constant due
to caching.

Future problem that we could run into, is peculiar characters being
encoded into the symbol name, leading to Z3 chocking or worse, undefined
behavior.