Implement casting from int to ptr
This is an extremely expensive operation - we have to consider all
possible pointers, and all possible offsets into them. Which means that
every time you cast from an int to a pointer, you have a huge clause,
the size of the number of pointers there are.

There's no real way around this when converting back from some abstract
address space into a pointer. The good solution, is to just not damn
well use such horrible pointer arithmatic.