Handle invalid-object irep during symex
I believe what goes on with this, is that we check in the __ESBMC_alloc
array to see whether or not the object has been free / invalidated /
whatever, which would lead to a false being written into the
__ESBMC_alloc array at its index.

Unfortunately, we can't implement this in Z3 because we don't know at
what point in time / SSA we're supposed to consider. So instead, when
we're replacing all dynamic obj ops, insert this tests against the value
of __ESBMC_alloc.

Requires some juggling because there are three functions called
replace_dynamic_allocation, something that sounds pretty war-crime like.