Correct symbol conversion
Previously it turns out we were taking a symbol and then updating is
fields with the correct value, instead of initializing a tuple then
binding it to the symbol. This leads to the actual symbol itself always
being a free variable.

It isn't obvious if this actually /breaks/ things, because all accesses
to the symbol (should) go through z3_convt::convert_identifier_pointer,
but it was certainly leading to insufficient data being in the
backtrace/output/whatever