Fix naming of addrspace struct type
This turns out to be a serious problem - because the addrspace array has
a different type between where we declare it and where we access it; for
some reason Z3 decides that they're /different objects/, and generates
another free variable of the same name (?).

Fix this up by using the same type name in these declarations.