Don't allow address range of an obj to wrap around
IE: if we're in bitvector mode, and allocate the start of a pointers
address space to be at the end of the address space, then the /end/ of
that pointer wraps around. Which is bad, and leads to incorrect pointer
reasoning later.

Fix this by asserting that the end of the pointers address range is
greater than the start of it.

(also, 1500GET)