Add a debug routine for z3 formula
The current problem is that we can have a couple of pages of formula
making up a single assignment or other thing that gets asserted, which
means that the contents of said formula have no associated names. This
further means that we can't get their values out from the core generated
by Z3.

So, this routine allows you to name a piece of AST and have it
automagically numbered too. So one can then add a label and then inspect
the insides of some larger formula via means of all these extra names.
Woo.