Eliminate some form of caching reliance
Possibly in the past it was always the case that these objects were in
the expression cache but it certainly isn't the case now. And if they
are, convert_bv is going to retrieve them from the cache.

So, always generate the correct pieces of AST instead of bailing out and
returning false