Fix a parser failure I'd inadvertantly introduced
When refactoring the C parsing goo it seems we've reached a situation
where "nil" can become part of a type. Which is fine. Except that then
it leads to errors where types aren't allowed to have qualifiers, but
they do have "nil".

So fix this by not considering nil as an "other" type.