Minor work around for change in behaviour
So, one of lucas' workarounds in the Z3 backend is ditching typecasts of
null, however he determines it's null by checking that it has no
operands. And it turns out that, from whatever I did to the ansi-c
frontend, NULL now translates into an irep with value=null /and/ with a
zero operand, which it didn't before.

So, modify this behaviour to match what Lucas' code expects.

Insert here complaints about strings.