Adjust pointer obj layout arrangement
So, limiting the available address range from 2^32 down to 15 for one
benchmark reduced the runtime from 15s to 0.15s (indeterminate amount of
Z3 in the latter). So clearly there's some benefit from restricting how
Z3 lays these objects out in memory.

So, switch to an arrangement where we chain all memory items so that
each item immediately follows the former. This instantly means that
there are a set of bugs that we fail to catch, but, they only ever occur
in the presence of extreme pointer arithmatic (ie, where there are
conditions that depend on the order of items on the heap / whatever,
_and_ there's pointer-jumping (add a number to a ptr, it goes out of
bounds, but always points into a valid object).

This'll get written up in documentation.