Optimise new memory layout
See previous commit for detail on memory layout; this change asserts the
fact that all addresses we deal with lie at the bottom of the address
space, and so Z3 doesn't need to consider addresses past that.

Assert this fact, allowing Z3 to only consider a much smaller range of
addresses when deciding things.