Make formula assertion more rigid
It's necessary for getting an unsat core, to make all the assumptions
we use literals, ie, not bits of formulas. Otherwise Z3 spits it out.

This requires making all formulas into literals, managed through
assert_formula, or if there's already a literal, we assert it's relation
to whatever it relates to and store the literal by itself.