Neuter pointer layout functionality
As opposed to the previous arrangement for allocating pointer objects
around the address space, this mechanism just gives everything a fixed
address. This satisfies the problem of needing an address space but not
knowing how to allocate it...

of course the problem is that then this doesn't catch conditions where
the pointer locations can actually be different, and some utterly
convoluted piece of code will pass verification. So, the workaround for
this is that the user passes in some additional optio to ESBMC
(unwritten yet), and we perform underapproximatio and widenning using
unsat cores to relax the fixed memory locations.

This balances out the overhead of leaving Z3 to explore all memory
layouts, with the problem to making the location allocation too fixed.