Enable the Z3 "preprocessor" in normal checking
There's no real documentation of what this option does, only that it
performs solving during hte preprocessing step. pretty cryptic.

However the observed effects are that no unsat core is generated if this
option is set to true, and verification takes for ever if it's turned
off, in different tests. So, if we're not going for an unsat core, we
always turn the option on to get the performance gain.