Define a "depend" rule for generating dependancies
These get generated in the normal fashion, from gcc, but are now placed
in the .deps file (for much cleaner directory listings). Also, it means
dependancy lists can be generated prior to compilation, which may fix
some of the parallel dependancy fails I've experienced.

The cpp makefile is patched as lex.yy.cpp isn't actually a file we have
when we make depenancy calculations (and its dependancies should be
defined in-file).