Seperate some heap liveness properties
It's increadibly unclear (and 100% undocumented) what such things as
"__ESBMC_alloc[]" and suchlike are for. As it turns out, both the
"valid_object" irep operation, and the "deallocated_object" irep
operation both use the value of __ESBMC_alloc[obj_identifier] as a bool
indicating the operations truth.

Clearly it's invalid, because if the flag in that array is true, then
both the valid_object and deallocated_object operations are true, which
should be mutually exclusive. Which is bad.

So, have an array to hold a bit representing whether the item has been
deallocated. As for the purpose of __ESBMC_alloc right now, its anyones
guess.

There's no requirement to update any "free" code so that it updates the
deallocated array; it makes an assignment to the "deallocated_object"
irep operation already, so that's translated to an assignment to the
__ESBMC_deallocated array later.

(It's yet again increadibly unclear how that happens, because the code
to do it exists in goto-programs/builtin_functions.cpp, rather than
goto-symex/symex_other.cpp or goto-symex/builtin_functions.cpp)