Eliminate "bl" option
It doesn't do anything. In fact, the only place it's read from the
options is when selecting whether to print "bit-vector" or
"integer/real" when solving, which just leads to the wrong fact being
printed.

As far as I can tell, it's extremely closely related to the int-encoding
option.

This fixes the 03_nec13 regression test.