Fix regress. test for simplification change
Due to a change in simplification that was leading to some bugs occuring
(c2ddba3c0), some regression tests now can't automatically determine the
unwind bound for the test. For example, this one, which involves a
subtraction.

I figure it's better to shoot for overall correctness rather than for
making this easier to use; so I'll be updating regression tests rather
than reverting the change to the simplifier.