Move ex_state initialization out of static inits
I've updated my machine to Fedora 16 - with a new version of gcc,
libstdc++ and suchlike. Unfortunately this leads to freshly compiled
copies of ESBMC crashing, which is bad. After some analysis, this would
appear to be due to C++'s static member initialization trainwreck.

The trouble is that the expression id to serialise-function mapping in
execution_statet is initialized before the string container hash maps,
and either because the hash map has changed or the initialization order
is now different due to the compiler, this now crashes.

So: I've fixed this by hoisting the expr id map out of static
initializers and into the execution_statet constructor. This is no
longer thread safe, but meh.

Insert here - repeated cursing of C++.