Reset state of cpp directory
This resets the state of the CPP directory to how it is in the master
branch right now - i.e., without any of the irep changes or the
problems of the cpp dir being deleted in the irep branch affecting it.

I know it's identical to master, because this commit is hacked together
to make the tree hash for that dir the same.

(The reason for this is because several changes haven't been brought in
by the previous merge - the assertion fixes for example).