Ensure there's a release note entry for a release
This will refuse to compile a release unless there's an entry in
release-notes.txt corresponding to this release name. Which is good,
because we won't end up releasing without an entry describing it.

On the other hand this means you can't compile a spurious release for
someone; something to be fixed in the future.