Permit unary-'s for signed types
The previous commit (c2ddba3c) that blocked these was a little too
broad: unary-'s will work when inverted (somewhat) with integers that
are classified as signed, as binary2integer will notice the signedness
and return an appropriate value.

This doesn't quite work with /huge/ integer constants; but that's the
fault of the internal representation.