Start making from-bv funcs; firstly, bool from bv
This depends heavily on the assumption that booleans are going to be
represented in a byte representation by an 8 bit integer, which as
mentioned in patch, should be configurable from somewhere else.

Presumably it can be another size too. Or, true and false can be
different values. Or, should be able to, because for now it's hacked
like this.