Make byte_extract... return correct sized bv
So as it turns out, byte extract means extract from the byte
representation, not meaning that it's only a single byte we're removing.

Note that this patch now makes this irep agnostic about the byte order.
IMO this is the way that it should be, the byte order is a feature of
converting integers into a bitvector, not the the more general "byte
extract" case. So I'll be implementing that to make to_bv and from_bv
perform the correct operation.

I don't think it's possible to have byte_extract have a different
endianness to the machine endianness we're simulating. If it is, I shall
be personally offended.