Make byte_update update correct sized bv
i.e., the width of the expr to insert, rather than just a byte.

As in 05e213f5, this makes the irep agnostic about the byte order it's
converting to/from. And as before, that'll be handled at integer
conversion time, not irep implementation.

(Right now I don't know what byte order Z3 uses internally).