Fix access to byte_ext.. manipulated piece of data
The problem is that in the memory model code, the dereference facility
has only the output of the points-to set. Which gives the _value_ that a
pointer points to, so an element of an array perhaps, or a struct
member.

However, we can't properly perform byte extracts with that value. We
need the full data entity. So pull the real array/struct/thing out of
the index/member expression.

Two things that are unresolved: nested arrays and suchlike will probably
need repeated fetching of the real array/struct/blah. And second,
whether or not the offset introduced by the index/member needs to be
taken into account (probably yes).

Still, working towards accuracy.