Further adjust symbolic-extract goo, for types.
The two operands to shr need to be the same width; therefore we need to
extend the length towards the size of the data bitvector.

It also turns out we don't need the elaborate operations in c162c346, as
imo anything that'll cause an integer overflow is going to cause a
bounds violation elsewhere.