Seriously change meaning of byte_update.
Previously byte_update had a value of the updated byte. However this has
lead to some problems. As mentioned in previous commits, the source data
for byte_extract is an element/member of a struct/array, not the
struct/array itself, due to the nature of the pointer analysis.

Which leads us onto this code: an assignment to a byte extract is
translated to a byte_update with almost the same operands, the result of
which is assigned to the original byte_extract's 1st operand.

And the problem: we can't update /past/ the member/index identified in that
first operand. If it's a byte, we can only assign a new byte. So a 32
bit write to an array of characters will perform a 32 bit extract on
something, which is then assigned to this 8 bit element that the
points-to analysis returned. Wheras in reality we want to update the
whole array or struct, because this byte update doesn't apply to just
one part of it.

The solution: change the meaning of byte update. Instead of updating a
byte or region in a particular subportion of a heap chunk or lump of
data, it updates the whole damn thing. This neatly eliminates any
worrying about what part is being updated at the symex layer, it's the
solvers problem.