Fix serious fallout from 85d399be
In 85d399be I updated the byte_update ireps to have the type of the base
object being operated upon - so that it could update a whole array or
struct, if the byte_update overran a particular element.

However, this still left the first operand of byte_update as an index or
member operation, with later code fishing out the base object from that.
Unfortunately, this can be constant propagated to the _value_ of that
element/member/whatever, rather than a reference to it. Which is bad.

(The specific case is that byte_update ends up assigning a value to
zero, because a character array reference got constant-propagated to the
value of the array element, zero).

Fix this by putting the reference to the base object in the first
operand of byte_update. There shouldn't be a change in solver behaviour
because it already fishes that out, but this way it won't be simplified
away.

(It still remains to be seen whether or not the offset into the object
introduced by the wrapper around it will affect program analysis later
on).