Byte updates of integers need an extra byte swap.
This is for circumstances where we need to write an integer into some
kind of array perhaps; the last time that we know that it's an integer
is when we convert it to a bitvector for the update, so it requires byte
swapping then. Hence, to_bv it.

(This will also fix the future bug of updating a piece of data with the
 value of a whole struct, which needs to be to_bv'd too).