Start hacking build scripts to use MSVC
We have to build under cygwin - gcc is required to generate dependancy
files for make. Additional pain comes from the fact that there are two
kinds of path: the shell will only accept cygwin paths, and MSVC will
only accept windows paths, so there's going to have to be some strict
seperation going on.

The plan here is to hack in MSVC support, then fix it to work on linux,
then flip back and forth until there's a system that works on both. Then
make it pretty. And then it'll be the heat-death of the universe anyway.