Adjust cpp build rules
Only build within the cpp makefile up to the point where we can hand
over to the other build rules. That way, we don't need to think about
gcc/msvc in that makefile.

This seems to expose the race condition I've experienced in the past
where y.tab.h isn't build in time, so I should be able to fix that now.