Begin beating irep serialisation portability
As documented in trac ticket #63, irep serialisation is a victim of
nonportable code using GNU-only facilities.

In cef36884 I removed what appeared to be a superfluous template
parameter... however as it turned out, it was a required feature
preventing ireps in the hash map from colliding. Without it, some ireps
aren't written out to the serialised file due to collisions, leading to
an unparseable file.

I really wish I hadn't devalued the term "war crime" in reference to
code so much.

Anyway, this patch begins moving towards a portable (but probably slow)
way of doing the same thing.