Generally hack around building 64 bit mingw
This is because on fedora at least (my machine) there aren't any
packages for installing the mingw32 windows 64 targetting toolchain.
Which is fine, I can just download that toolchain myself, but it doesn't
get the same layout and configuration as the package does.

This patch adds all the hacks required for just taking that toolchain
and putting it in /usr. Libraries are in different places, and the
system root is in a different place, but meh.