Hack at solvers dependancy situation.
If I "touch solvers/z3/z3_conv.cpp", and run make, then z3.o will be
rebuilt, but not solvers.o. It takes until the next run to work that
out.

This seems to be because make determines that the object files that
solvers.o depends on are not updated _before_ the phony directory-makes
are run (and as a result files modified).

So, plug the directory-make situation into a $shell so that it's run
immediately when the makefile is parsed and evaluated. That way, the
file update checks are always done after make is run for each dir. Woo.