Fix nonrecent union access
If we access a union member, we return the most recent member assigned -
because it's overwritten everything else. However, that type might be
anything. This leads to a mismatch between the expr type and Z3 type
being used, which is bad.

For some reason this only exhibits itself on windows? Ah, because
something causes it to run in BV mode.