Ensure build id in ESBMC is always up to date
Previous situation, buildidobj.o would be built and then never updated
after that; so someone is quite free to modify the tree, rebuild, and
the ID won't notice that the tree is dirty.

This way, whenever esbmc is triggered to be relinked, we end up
regenerating that object and ID string.