Fixup bodged solvers directory make rules
Rather than stuffing calls to make in $(shell...), instead make the
default rule remake directories (in a _new_ make instance), then make
solvers.o in yet another make instance.

This gets around the problem that make won't see updates to files that
are performed by a submake process. So if z3.o is rebuilt by a lower
process, we don't know about it.