Discard previously used if guards
It appears up until now, whenever one takes a branch, CBMC symex code
takes the current if guard, i.e. the guard on this current line of
execution, adds it to the guard from the current branch, and pushes this
new branch onto a stack of if guards. The topmost guard represents the
guard of the path we're taking from this current branch.

However, the data in that stack of guards is never accessed again. Only
the top value is ever accessed, if another conditional branch is taken.
So, all that data is hanging around doing absolutely nothing, using up
memory.

It really seems we don't need all this additional data; so only store
that top value. For the 12_lms_new benchmark this reduces peak memory
usage by about 10x. It also appears the total memory usage of that data
contributed about 2Gb, which was about 99% of memory during symex
destruction.