Define method to consolidate current if guard
We have the additional problem that every time cur_if_guard is updated,
it accumulates. So if we have some kind of loop unwound 200 times, and
the guard is instanciated each time, we'll get something in the order of
O(n^2) complexity. This uses memory, but more importantly  I think when
it gets to the SMT solver it'll solve each condition AST individually,
which is vastly inefficient.

So - the plan is to, every time cur_if_guard is updated, assign to a new
guard variable. This prevents accumulation, should also prevent each
guard being evaulated seperately, as speculated above.