Absolutely thrash z3_capi
This causes the z3 context pointer to be stored in the z3_capi object,
and to actually be valid and used. This saves us an extra parameter
setup on each invocation of a capi function

This also led to some "display" functions being deleted because nothing
uses them and they're to uninteresting to maintain. Bulk changes also
occured to z3_conv, as it uses a bunch of the changed z3_capi methods.