Fix bad behaviour of extracted array values
As documented in the comments, just pulling data out and sticking it in
a map doesn't give us the correct results - the same element in the
array can have different values assigned in the course of the arrays
lifetime, and we need to have the latest value.

So, pull all data out, feed into map in reverse (duplicate entries will
be overwritten by the latest copy), and then dump into array expr
operands.

It remains to be seen how unassigned array elements get represented
here. They probably don't.