Eliminate START_THREAD instruction
The plan being to replace it in favour of ESBMC intrinsics. This
simplifies conversion and analysis over code.

However there's a complication - pointer analysis might not work,
previously it might have flowed over the START_THREAD instruction and
successfully collected pointer info to allow the function pointer or
args to be dereferenced. Or, it might not.

If it turns out to be a problem we'll just end up dereferencing the
pointer when it's set.