Start generating funcptr deref infrastructure
The function defined in this patch fetches a list of function symbols
that exist in the expr we're dealing with, that can be called from this
function pointer. It also collects their guards.

Once that data is collected, the next move is to call each one and then
merge them in the future.