Take a different approach to storing funcall data
Instead of keeping a lot of our data hanging around explicitly, instead
generate a bunch of state merges at the starts of functions that we'll
be calling. This means that to setup a function call we only need to:

1) set the state guard to false
2) jump to the start of a function

And the original state will be merged into place and the function run.