Move towards reusing symex_function_call_code
there's a bunch of stuff in there that could be reused, most notably
setting up a new frame and copying arguments, some noninsignificant
actions.

So, move towards being able to reuse it when setting up the function
calls for a function pointer dereference. We'll force the state merge to
happen manually, then jump to the original function call location, then
invoke the function call and pretend it's a normal one.