Remove function call type checking situation.
There are _comments_ stuffed into the type, for example the location
that it was declared. This means some things that are compatible are
compared as being different, because they were declared in different
places.

Ewww.

Removing this, we're leaving it to other goto symex routines to check
that function pointer types match. Which hopefully they do.