Remove an invocation of "symex"
This "symex" function appears to be extremely old, and repeatedly
executes instructions rather than arranging them through symex_step.

Clearly this is largely incompatible with the existing way of doing
symex. It's also large and unwieldy, so I'd like to be rid of it.

So we remove this invocation when dealing with "cpp-try". From some
analysis, it seems this irep never reaches this point, as all the
current cpp tests don't change behaviour once I removed this.
Additionally it seems that goto convert code eliminates this irep
completely.