Move symex responsibility to execution_statet
So, art is no longer a goto_symext subclass, and instead passes all of
the symex stuff down to the currently executing execution_state.

The next step is to seperate these two things, so that art is only
responsible for storing and beating scheduling of states, and
execution_statet does such things as working out when a context switch
needs to occur, or whatever.