Make ex_state and goto_symext use same "target".
They've always (ish) been the same thing, but only due to hacks. The
most notable feature of this change is the lifetime of the target
object: we pass in a symex_target_equationt from bmc.h, and that gets
repeatedly cloned every time we make a new execution_statet.

This mean that the concrete class of the target isn't known at all by
the goto-symex directory.