Rename _DFS_traversed, remove _exprs
Didn't intend to squash these two things into the same commit, but never
mind.

_exprs removed because, well, I can't see anywhere where this actually
affects the flow of the program. Things are put into the exprs vector,
representing the most recently analysed expr or something, but then
aren't considered by anything.

Only read I can see is that it's put into last_global_expr upon
occasion, and upon occasion that's used to work out if we shouldn't
consider an ileave due to POR.

It's difficult to say whether or not this is going to lead to more
ileaves being deconsidered due to POR. So I'll say this: because it's
100% unclear, ditch it. It can only be bad.