Remove a final use of is_dynamic_object
We already mark the pointer as dynamic in the goto-symex version of
builtin functions, so this is unneeded.

It's quite worrying, as to whether or not the guard assignment statment
_does_ anything. At runtime I can't see either the dynamic obj or
neg_valid_expr expr turn up in t_a->guard. I have the very nasty
suspicion that, whatever that expression is supposed to do in terms of
operator overloading, what's actually happening is the comma operator
evaluating and discarding the first to exprs, then assigning the final
to t_a->guard.

Let sleeping dogs lie, though.