Revert various object munging commits
This reverts the plan to seperate the renaming process and the renaming
data; on second thoughts, the only real thing that's desired is that
there's _some_ way of linking back to execution_statet for level1 and
level2t objects.

This is best achived by extending level1t/level2t, and cloning it
repeatedly.