Update ex_state references in goto_statets on copy
The point being that in all level2 objects, there's a fresh and up to
date reference to th ex_state that owns it. Unfortunately there's pretty
much no mechanism to have these automatically updated, or otherwise to
have new level2 objs have access to the new ex_state.

(Ish. It can be hacked. But lets see how unperformant this is before
hacking it).