merge_goto before ex_state's symex_step.
Previously this wasn't happening for instructions that were interpreted
by ex_state's symex_step. Which is bad.

It also seems there's no decent way of preventing merge_goto from being
called twice, without putting it somewhere not in symex_step. We can't
put it at the _end_ of symex_step, as we get some confusing segfaults,
and the only other way is to pre-decide whether the insn is going to go
down goto_symext::symex_step, or the ex_state version, and decide to
call merge_gotos based on that.

This current approach is probably the simplest; we can worry about
efficiency later.