Move cswitch decision from symex_step to assign
The point here being that we're ending up with a situation whereby
execution_statet is now going to be responsible for deciding anything
regarding context switches, removing complexity from the base symex goo.

Note that this change is engineered for performing the assignment, then
possibly cswitching, after it. This is different from the other
instructions interpreted, which fully abort before actually incrementing
the program counter. Exactly why this is the case isn't documented or
clear.

Also note that this is going to break down and cry if multiple
assignments or other cswitch-generating events are caused in one
instruction, as there's no definition about which ones are repeated or
what order they occur in. To which the solution is to /not/ allow
multiple cswitching events in a single instruction. Which imo is 100%
reasonable.