Switch all instructions to cswich after operation
I.E., any context switch that is going to happen, will occur _after_ the
particular operation being performed. See below for a more general why.

I believe this is generally equivalent to what we're already doing;
essentially, all operations themselves are atomic. And context switches
are put between them.

Now that we have a situation where there's a context switch point at the
start of each thread, and one at the end on thread/main termination, all
that we require is that there's a context switch _somewhere_ in the
middle, between each visible instruction. Whether it's before or after
is now immaterial.

Picking making context switches _after_ operations makes the most sense.
It means we don't need to reconsider reexecuting instructions (below).

This patch causes no change in behaviour in the regression tests, the
same number succeed with and without.

~

So why this direction? Observe the reexecute_instruction flag. Observe
that it's set whenever we reach a context switch point. Observe that it
causes an instruction we didn't execute (because we cswitched before it)
to re-create another context switch point (infinitely recursively).

Now observe that if we context switch again, with reexecute_instruction
set, then a future random context switch point is going to be clobbered
by this flag, whether or not it should cause a context switch. This is
bad.

Solution: have all instructions cause context switches _after_ they
execute, and then nuke the reexecute_instruction feature.